Refactor compat package
